+++ tests/Test/StorageApiTestCase.php
@@ -17,6 +17,7 @@ class StorageApiTestCase extends \PHPUnit_Framework_TestCase
 	public function setUp()
 	{
 		$this->_client = new Keboola\StorageApi\Client(STORAGE_API_TOKEN, STORAGE_API_URL);
+		$this->_client->setBackoffMaxTries(1);
 	}
 
 	/**